Its completed given task

+++ DotNetTask/Controllers/ProgramDataController.cs
@@ -27,6 +27,10 @@ namespace DotNetTask.Controllers
         [HttpGet]
         public ActionResult GetQuestionTypes()
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
             var values = Enum.GetValues(typeof(QuestionTypes))
                              .Cast<QuestionTypes>()
                              .Select(q => new { Name = q.ToString(), Value = (int)q })
@@ -44,6 +48,10 @@ namespace DotNetTask.Controllers
         [HttpPost("SaveProgram")]
         public async Task<ActionResult> SaveProgram(ProgramDataRequestDTO programRequest)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
             ProgramData data = programRequest.ToProgramsDataModel();
             ProgramData programResponse = await _programDataRepository.CreateProgramDataAsync(data);
             return Ok(programResponse);
@@ -58,7 +66,11 @@ namespace DotNetTask.Controllers
         [HttpPut("UpdateQuiz/{programId}/{questionId}")]
         public async Task<ActionResult> UpdateQuestion(QuestionDataDTO quiz, string programId, string questionId)
         {
-           
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
+
             var result = await _programDataRepository.UpdateQuestionsAsync(quiz, programId, questionId);
             if (result == null)
             {
@@ -75,7 +87,32 @@ namespace DotNetTask.Controllers
         [HttpDelete("DeleteQuiz/{programId}/{questionId}")]
         public async Task<ActionResult> DeleteQuestion(string programId, string questionId)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
             var result = await _programDataRepository.DeleteQuestionsAsync(programId, questionId);
+            if (result == null)
+            {
+                return NotFound();
+            }
+            return Ok(result);
+        }
+
+
+        /* 
+         * Get program data
+         * @param programId
+         */
+
+        [HttpGet("GetProgram/{programId}")]
+        public async Task<ActionResult> GetProgramData(string programId)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
+            var result = await _programDataRepository.GetProgramAsync(programId);
             if (result == null)
             {
                 return NotFound();
